Forslag til dataoverføring
Ryddet i filene og lagt til et forslag til kommunikasjon

+++ SmartKollektiv_ESP32_main_pio/src/main.cpp
@@ -44,7 +44,7 @@ WiFiClient espClient;
 PubSubClient client(espClient);
 long lastMsg = 0;
 char msg[50];
-int value = 0;
+long int value = 0;
 
 void setup_wifi() {
